requirements.txt
flask
ngrok
datetime
mysql-connector-python
pre-commit
paho-mqtt